add register and login

+++ src/main/java/com/hidear/law/core/log/factory/LogFactory.java
@@ -21,7 +21,7 @@ public class LogFactory {
      * @author fengshuonan
      * @Date 2017/3/30 18:45
      */
-    public static OperationLog createOperationLog(LogType logType, Long userId, String bussinessName, String clazzName, String methodName, String msg, LogSucceed succeed) {
+    public static OperationLog createOperationLog(LogType logType, Integer userId, String bussinessName, String clazzName, String methodName, String msg, LogSucceed succeed) {
         OperationLog operationLog = new OperationLog();
         operationLog.setLogtype(logType.getMessage());
         operationLog.setLogname(bussinessName);
@@ -40,7 +40,7 @@ public class LogFactory {
      * @author fengshuonan
      * @Date 2017/3/30 18:46
      */
-    public static LoginLog createLoginLog(LogType logType, Long userId, String msg, String ip) {
+    public static LoginLog createLoginLog(LogType logType, Integer userId, String msg, String ip) {
         LoginLog loginLog = new LoginLog();
         loginLog.setLogname(logType.getMessage());
         loginLog.setUserid(userId);
