+++ src/main/java/com/hidear/law/core/log/factory/LogTaskFactory.java
@@ -25,7 +25,7 @@ public class LogTaskFactory {
     private static LoginLogRepository loginLogRepository = Db.getRepository(LoginLogRepository.class);
     private static OperationRepository operationRepository = Db.getRepository(OperationRepository.class);
 
-    public static TimerTask loginLog(final Long userId, final String ip) {
+    public static TimerTask loginLog(final Integer userId, final String ip) {
         return new TimerTask() {
             @Override
             public void run() {
@@ -54,7 +54,7 @@ public class LogTaskFactory {
         };
     }
 
-    public static TimerTask exitLog(final Long userId, final String ip) {
+    public static TimerTask exitLog(final Integer userId, final String ip) {
         return new TimerTask() {
             @Override
             public void run() {
@@ -83,7 +83,7 @@ public class LogTaskFactory {
 //        };
 //    }
 //
-    public static TimerTask exceptionLog(final Long userId, final Exception exception) {
+    public static TimerTask exceptionLog(final Integer userId, final Exception exception) {
         return new TimerTask() {
             @Override
             public void run() {
