+++ src/main/java/com/hidear/law/modular/dao/LoginLogRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by Administrator on 2017/7/8.
  */
-public interface LoginLogRepository extends JpaRepository<LoginLog,Long>{
+public interface LoginLogRepository extends JpaRepository<LoginLog,Integer>{
 }
