+++ src/main/java/com/hidear/law/modular/dao/OperationRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by acer on 2017/7/8.
  */
-public interface OperationRepository extends JpaRepository<OperationLog,Long>{
+public interface OperationRepository extends JpaRepository<OperationLog,Integer>{
 }
